+++ TableView/BaseControls/QAAttributedLabel/BaseClasses/Category/NSMutableAttributedString+QAAttributedString.m
@@ -172,7 +172,7 @@
     
     unsigned int count = 0;
     objc_property_t *properties = class_copyPropertyList([NSMutableAttributedString class], &count);
-    for (int i = 0; i < count; i++) {
+    for (NSUInteger i = 0; i < count; i++) {
         objc_property_t property = properties[i];
         const char *name = property_getName(property);
         NSString *key = [NSString stringWithUTF8String:name];
